+++ htdocs/ecotaxdeee/core/modules/modEcoTaxDeee.class.php
@@ -61,7 +61,7 @@ class modEcoTaxDeee extends DolibarrModules
         $this->editor_name = 'NLTechno';
         $this->editor_url = 'https://www.nltechno.com';
 		// Possible values for version are: 'development', 'experimental', 'dolibarr' or version
-		$this->version = '4.1.0';
+		$this->version = '4.1.1';
 		// Key used in llx_const table to save module status enabled/disabled
 		$this->const_name = 'MAIN_MODULE_'.strtoupper($this->name);
 		// Name of png file (without png) used for this module.