+++ htdocs/ecotaxdeee/ChangeLog.md
@@ -1,17 +1,23 @@
 # ChangeLog MODULE EXTERNALBACKUP FOR <a href="https://www.dolibarr.org">DOLIBARR ERP CRM</a>
 
 
+## 4.1.1 Unreleased
+
+- Use the VAT rate of the eco tax product instead of highest vat rate when a predefined product
+  is defined for the eco tax line.
+
+
 ## 4.1.0
 
-Compatibility with Dolibarr 7.0
+- Compatibility with Dolibarr 7.0
 
 
 ## 4.0.0
 
-Compatibility with Dolibarr 6.0
+- Compatibility with Dolibarr 6.0
 
 
 ## 1.0
 
-Initial version
+- Initial version